Updated parsing for more cases
The parsing now also catches `get` calls and handles edge cases properly.

Only remaining edge case not handled is when a user excapes the a quote in the var name rather
than using the alternate quote.
This case is ignored since running into it would mean the user is two-fold following bad practices:
- quotes in env var names and escaping quotes rather than
- using the alter quotes for handling interior quotes in a string.