Add support for configurable environments
Enable configurable environments based on expected
env vars from the notebook source code. This is
acomplished by identifying the vars from parsing the
notebook source, and asking for the user to input suggested
values to these vars wich is then used for preparing the
runtime environment before executing the notebook.

Closes #5